pushing db up to heroku

+++ db/migrate/20111210174043_create_contacts.rb
@@ -3,7 +3,7 @@ class CreateContacts < ActiveRecord::Migration
     create_table :contacts do |t|
       t.string :name
       t.string :email
-      t.integer :phone
+      t.bigint :phone
       t.string :address
       t.string :city
       t.string :state
